requirements.txt
# Core AutoGen and LLM dependencies
autogen-agentchat>=0.4.0
autogen-core>=0.4.0
autogen-ext[openai]>=0.4.0
openai>=1.0.0
anthropic>=0.20.0

# RAG and Vector Database
chromadb>=0.4.24
langchain>=0.1.0
langchain-community>=0.0.20
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Web Research Tools
requests>=2.31.0
beautifulsoup4>=4.12.2
selenium>=4.15.0
googlesearch-python>=1.2.3
newspaper3k>=0.2.8

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
openpyxl>=3.1.2
markdown>=3.5.1

# Data Analysis
pandas>=2.1.0
numpy>=1.25.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0

# Web Framework and UI
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
gradio>=4.7.0

# Memory and State Management
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Evaluation and Testing
deepeval>=0.20.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Safety and Monitoring
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
prometheus-client>=0.19.0
structlog>=23.2.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
click>=8.1.0
rich>=13.7.0
tqdm>=4.66.0
python-multipart>=0.0.6

# Development Tools
black>=23.10.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.5.0
